Use submodules and  moved to subdir.
