--- about.md
@@ -1,6 +1,0 @@
----
-title: About
-layout: single
----
-
-The IOOS's notebook gallery is a collection of tutorials and examples of the many technologies and data sources available to the community.
