--- index.md
@@ -1,14 +1,0 @@
----
-title: IOOS's Notebook Gallery
-layout: single
----
-
-<div id="gallery">
-<div><figcaption>Fetching data from CO-OPS SOS</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/2016-10-12-fetching_data"><img src="images/2016-10-12-fetching_data.png"></a></div>
-<div><figcaption>CF, UGRID, and SGRID Conventions</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/CF-UGRID-SGRID-conventions"><img src="images/CF-UGRID-SGRID-conventions.png"></a></div>
-<div><figcaption>Using Siphon to query the NetCDF Subset Service</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/SIPHON_NCSS_HFR_DataAccess"><img src="images/SIPHON_NCSS_HFR_DataAccess.png"></a></div>
-<div><figcaption>Demonstrate CSW query</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/csw_search_example"><img src="images/placeholder.png"></a></div>
-<div><figcaption>Glider RU22</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/glider_data_example"><img src="images/glider_data_example.png"></a></div>
-<div><figcaption>HYCOM-glider</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/glider_hycom_iris"><img src="images/glider_hycom_iris.png"></a></div>
-<div><figcaption>siphon-ncss</figcaption><a href="{{ site.url }}{{ site.baseurl }}/notebooks/siphon-ncss"><img src="images/siphon-ncss.png"></a></div>
-</div>
