+++ webpage/_config_dev.yml
@@ -1,0 +1,12 @@
+# Welcome to Jekyll!
+#
+# This config file is meant for settings that affect your entire site, values
+# which you are expected to set up once and rarely need to edit after that.
+# For technical reasons, this file is *NOT* reloaded automatically when you use
+# `jekyll serve`. If you change this file, please restart the server process.
+
+# Site Settings: only include setting for development on 'localhost' to override
+# GitHub settings:
+
+url                      : "http://localhost:4000"
+baseurl                  : "/notebooks_demos"
